Obsługa kątów jako float, poprawa geometrii i logów
Zmieniono typy kątów z int na float dla większej precyzji obliczeń. Ulepszono funkcję StronaSzukaj, aby dokładniej określała stronę na podstawie kąta i położenia linii. Zaktualizowano wywołania funkcji związanych z sąsiadami i nakładaniem się linii. Tekst na profilach jest teraz obracany zgodnie z rzeczywistym kątem. Dodano szczegółowe logi diagnostyczne do debugowania geometrii. Zwiększono domyślną szerokość okna w Generatorze oraz poprawiono obsługę nietypowych przypadków (np. trójkąty, styki T3/T3). Usunięto stare, zakomentowane logi. Zmiany poprawiają precyzję, wizualizację i ułatwiają debugowanie.

+++ Shared/ViewModels/DaneKwadratu.cs
@@ -9,7 +9,7 @@ namespace GEORGE.Shared.ViewModels
         public Guid RowIdSasiadaStronaA { get; set; } = Guid.Empty; //id elementu w bazie danych sąsiad strona A przecinający (góra/lewa)
         public Guid RowIdSasiadaStronaB { get; set; } = Guid.Empty; //id elementu w bazie danych sąsiad strona B przecinający (dół/prawa)
         public string? RowIdRegionuSasiada { get; set; } //id elementu w bazie danych
-        public int KatLinii { get; set; } //kierunek linii w stopniach
+        public float KatLinii { get; set; } //kierunek linii w stopniach
         public string? Strona { get; set; } //kierunek linii w stopniach
         public double OffsetTop { get; set; } = 0; //odległość linii od górnej krawędzi elementu pobrana z bazy danych
         public double OffsetBottom { get; set; } = 0;
